feat: auth route guard
