+++ apps/api/src/routes.ts
@@ -1,12 +1,13 @@
 import { otel } from "@hono/otel";
-import { type auth, authHandler, requireAuth } from "@paperjet/auth";
+import { type auth, authHandler, requireAdmin, requireAuth } from "@paperjet/auth";
 import { envVars, logger } from "@paperjet/shared";
 import { Hono } from "hono";
 import { serveStatic } from "hono/bun";
 import { logger as honoLogger } from "hono/logger";
 import { poweredBy } from "hono/powered-by";
 import { corsMiddleware } from "./lib/cors";
-import { type AdminRoutes, adminRouter } from "./routes/admin";
+import { type AdminRoutes, v1AdminRouter } from "./routes/admin";
+import { type InternalRoutes, internalRouter } from "./routes/internal";
 import { type ApiKeysRoutes, v1ApiKeyRouter } from "./routes/v1/api-keys";
 import { type ExecutionRoutes, v1ExecutionRouter } from "./routes/v1/executions";
 import { v1WorkflowRouter, type WorkflowRoutes } from "./routes/v1/workflows";
@@ -27,6 +28,7 @@ app.use(
 );
 app.use("/api/*", corsMiddleware);
 app.use("/api/v1/*", requireAuth);
+app.use("/api/v1/admin/*", requireAdmin);
 app.on(["POST", "GET"], "/api/auth/*", authHandler);
 
 app.get("/api/health", async (c) => {
@@ -38,7 +40,8 @@ app.get("/api/health", async (c) => {
 
 app
   .basePath("/api")
-  .route("/admin", adminRouter)
+  .route("/internal", internalRouter)
+  .route("/v1/admin", v1AdminRouter)
   .route("/v1/api-keys", v1ApiKeyRouter)
   .route("/v1/workflows", v1WorkflowRouter)
   .route("/v1/executions", v1ExecutionRouter);
@@ -59,3 +62,4 @@ export type { ApiKeysRoutes };
 export type { ExecutionRoutes };
 export type { WorkflowRoutes };
 export type { AdminRoutes };
+export type { InternalRoutes };
