+++ apps/dashboard/src/queries/server-info.tsx
@@ -1,15 +1,15 @@
-import type { AdminRoutes } from "@paperjet/api/routes";
+import type { InternalRoutes } from "@paperjet/api/routes";
 import { queryOptions } from "@tanstack/react-query";
 import { hc } from "hono/client";
 
-const adminClient = hc<AdminRoutes>("/api/admin");
+const internalClient = hc<InternalRoutes>("/api/internal");
 
 export const serverInfoQueries = {
   serverInfo: () =>
     queryOptions({
       queryKey: ["server-info"],
       queryFn: async () => {
-        const response = await adminClient["server-info"].$get({});
+        const response = await internalClient["server-info"].$get({});
         return response.json();
       },
       staleTime: 5000,
